fixing service and cleanup code
