+++ server/qtservice/qtservice.pri
@@ -1,4 +1,4 @@
-include(../common.pri)
+#include(../common.pri)
 INCLUDEPATH += $$PWD
 DEPENDPATH += $$PWD
 !win32:QT += network
